Generic SpliceTask added, DisneyAC converted and verification added, MMLC1 and MMXLC1 reimpl
