+++ src/gex/lib/tasks/impl/mmlc1/metadata.json
@@ -1,7 +1,7 @@
 {
     "in": {
         "files": {
-            "proteus": {
+            "source": {
                 "filename": "Proteus.exe",
                 "versions": {
                     "Steam": {
