+++ src/gex/lib/tasks/impl/mmxlc1/metadata.json
@@ -1,7 +1,7 @@
 {
     "in": {
         "files": {
-            "RXC1": {
+            "source": {
                 "filename": "RXC1.exe",
                 "versions": {
                     "Steam": {
